ml/requirements.txt
nltk==3.9.1
pandas==2.3.2
pip==25.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sentence-transformers==5.1.1
sentencepiece==0.2.1
torchvision==0.23.0
